- Move generic utility methods to "security/utils" module
