+++ geonode/base/api/serializers.py
@@ -37,8 +37,12 @@ from geonode.base.models import (
     SpatialRepresentationType,
     ThesaurusKeyword,
 )
-from geonode.base.utils import build_absolute_uri, get_resources_with_perms
-from geonode.groups.models import GroupCategory, GroupProfile
+from geonode.groups.models import (
+    GroupCategory,
+    GroupProfile)
+
+from geonode.base.utils import build_absolute_uri
+from geonode.security.utils import get_resources_with_perms
 
 import logging
